+++ docs/src/examples/components/Dialog/Variations/index.tsx
@@ -6,14 +6,9 @@ import ExampleSection from 'docs/src/components/ComponentDoc/ExampleSection'
 const DialogVariationsExamples = () => (
   <ExampleSection title="Variations">
     <ComponentExample
-      title="Backdrop"
-      description="A dialog can have a backdrop on its overlay."
-      examplePath="components/Dialog/Variations/DialogExampleBackdrop"
-    />
-    <ComponentExample
-      title="Inert"
-      description="Whether the dialog should be inert, e.g. not dismiss when focusing/clicking outside of the dialog. Hides an overlay."
-      examplePath="components/Dialog/Variations/DialogExampleInert"
+      title="Modal"
+      description="Whether content outside should be inert, e.g. when focusing/clicking outside is not possible and the dialog will have a visible overlay."
+      examplePath="components/Dialog/Variations/DialogExampleModal"
     />
   </ExampleSection>
 )